+++ include/encode/bitwriter.hpp
@@ -24,14 +24,14 @@ class BitWriter {
         auto write_bit(uint8_t bit) -> void;
         auto write_bits(uint64_t value, uint64_t n, std::endian endian = std::endian::big) -> void;
 
-        auto write_unary(uint8_t value, uint8_t bit) -> void;
-        auto write_unary_with_terminator(uint8_t value, uint8_t bit) -> void;
+        auto write_unary(uint64_t value, uint8_t bit) -> void;
+        auto write_unary_with_terminator(uint64_t value, uint8_t bit) -> void;
 
         auto write_gamma(uint64_t value) -> void;
         auto write_delta(uint64_t value) -> void;
         auto write_minimal_binary(uint64_t value, uint64_t z) -> void;
-        auto write_zeta(uint64_t value, uint8_t k) -> void;
-        auto write_golomb(uint64_t value, uint8_t b) -> void;
+        auto write_zeta(uint64_t value, uint64_t k) -> void;
+        auto write_golomb(uint64_t value, uint64_t b) -> void;
 
         auto flush() -> void;
 };
